Moves HTTPMethod into Request package.
