+++ src/main/java/request/HTTPMethod.java
@@ -1,4 +1,4 @@
-package controllers;
+package request;
 
 import java.util.ArrayList;
 import java.util.List;
