+++ src/test/java/controllers/HomePageTest.java
@@ -2,10 +2,11 @@ package controllers;
 
 import org.junit.Before;
 import org.junit.Test;
+import request.HTTPMethod;
 import request.HTTPRequest;
 import response.HTTPResponse;
 
-import static controllers.HTTPMethod.*;
+import static request.HTTPMethod.*;
 import static org.junit.Assert.*;
 import static response.StatusLine.OK;
